+++ examples/demo1-forms2.js
@@ -1,5 +1,11 @@
+/* eslint-disable react/no-string-refs */
+
 import React from 'react';
-import { NumberInput } from '../src';
+import {
+  NumberInput, isLte,
+  Button,
+  notify,
+} from '../src';
 import {
   ExampleLabel, exampleStyle,
 } from './demo1-common';
@@ -8,42 +14,79 @@ class FormExample extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      a: 3,
+      a: null,
       b: 4,
-      c: 5
     };
   }
 
   render() {
     return (
       <div style={exampleStyle}>
-        <ExampleLabel>Form example</ExampleLabel>
-        <NumberInput ref="a"
-          value={this.state.a}
-          onChange={this.onChange}
-          errors={['multiple', 'errors']}
-          required
-        />
-        <NumberInput ref="b"
-          value={this.state.b}
-          onChange={this.onChange}
-          required
-        />
-        <NumberInput ref="c"
-          value={this.state.c}
-          onChange={(ev, c) => this.setState({ c })}
-          required
-        />
+        <ExampleLabel>Uncontrolled vs. controlled form example</ExampleLabel>
+        {this.renderUncontrolled()}
+        {this.renderControlled()}
       </div>
     );
   }
 
-  onChange = async () => {
-    console.log('Changing...');
-    const a = this.refs.a.getValue();
-    const b = this.refs.b.getValue();
-    console.log({ a, b })
+  renderUncontrolled() {
+    const props = { required: true, validators: [isLte(10)] };
+    return (
+      <div>
+        <b>Uncontrolled:{' '}</b>
+        <NumberInput ref="a1" value={null} {...props} />
+        <NumberInput ref="b1" value={3} {...props} />
+        <Button onClick={this.onSubmit('1')}>Read values</Button>
+      </div>
+    );
+  }
+
+  renderControlled() {
+    const props = {
+      required: true,
+      validators: [isLte(10)],
+    };
+    return (
+      <div>
+        <b>Controlled:{' '}</b>
+        <NumberInput ref="a2"
+          value={this.state.a}
+          onChange={(ev, val) => this.setState({ a: val })}
+          {...props}
+        />
+        <NumberInput ref="b2"
+          value={this.state.b}
+          onChange={(ev, val) => this.setState({ b: val })}
+          {...props}
+        />
+        <Button onClick={this.onSubmit('2')}>Read values</Button>
+      </div>
+    );
+  }
+
+  onSubmit = (suffix) => async () => {
+    try {
+      const [a, b] = await Promise.all([
+        this.refs[`a${suffix}`].validateAndGetValue(),
+        this.refs[`b${suffix}`].validateAndGetValue(),
+      ]);
+      notify({
+        msg: `Values: ${a}, ${b}`,
+        icon: this.context.theme === 'mdl' ? 'thumb_up' : 'thumbs-up',
+      });
+    } catch (err) {
+      notify({
+        msg: 'Validation failed',
+        type: 'error',
+        icon: this.context.theme === 'mdl' ? 'warning' : 'bullhorn',
+      });
+    }
+  }
+
+  readValues = async () => {
   }
 }
+
+FormExample.contextTypes = { theme: React.PropTypes.any };
 
 export default FormExample;